+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-xxx
+0.5.32 / 2018-05-14
 ===================
 - Bugfix in `RequestList` - skip invalid in-progress entries when restoring state
+- Added `request.ignoreErrors` options. See documentation for more info.
 
 0.5.31 / 2018-05-11
 ===================